# Conflicts:
#   ProjectApp/db.py

+++ ProjectApp/elements/element_views.py
@@ -5,16 +5,7 @@ bp = Blueprint("element", __name__, url_prefix='/elements')
 
 @bp.route("/", methods = ["GET", "POST"])
 def show_elements():
-    form = ElementForm()
-    # ask teammates about what to make of new_element since element_orde and competency_id are no longers forms
-    if request.method == "POST" and form.validate_on_submit():
-        new_element = Element(None, form.element_order.data, form.element.data, form.element_criteria.data, form.competency_id.data)
-        try:
-            get_db().add_element(new_element)
-            elements, prev_page, next_page = get_db().get_elements(page_num=1, page_size=50)
-        except ValueError as e:
-            flash(e)
-    return render_template("elements.html", elements = elements, form = form)
+    return redirect(url_for("competency.show_competencies"))
 
 @bp.route("/<int:element_id>", methods = ["GET", "POST"])
 def show_element(element_id):
@@ -41,9 +32,13 @@ def show_element(element_id):
 @bp.route("/delete/<element_id>/")
 def delete_element(element_id):
     try:
+        element = get_db().get_element(int(element_id))
+        if element == None:
+            flash("element not found")
+            abort(404)
         get_db().delete_element(element_id)
         message = f'deleted element with id {element_id}'
         flash(message)
     except ValueError as e:
         flash(str(e))#there can be many reasons for this: element id doesnt exist or its the last of its competency
-    return redirect(url_for("element.show_elements"))
+    return redirect(url_for("competency.show_competency", id = element.competency_id))
